make imavid more modular
